madewithflyte/requirements.txt
flytekit>=1.5.0
scikit-learn
pandas
pyarrow
fastparquet
nltk
numpy
transformers
flytekitplugins-ray
openai
matplotlib
seaborn
tqdm
python-dotenv